Update feature/forget-password
Switch from hashtable to ConcurrentDictionary to guarantee threading

seperated utility class ResetCode to new file in Common folder

resolve miscellaneous warnings
