+++ FlashcardApp.Api/Dtos/UserDtos/ForgotPasswordDto.cs
@@ -4,6 +4,6 @@
     {
         [Required(ErrorMessage = "Please enter your email address.")]
         [EmailAddress(ErrorMessage = "The email address is not valid.")]
-        public string Email { get; set; }
+        public string Email { get; set; } = string.Empty;
     }
 }
