+++ FlashcardApp.Api/Program.cs
@@ -1,61 +1,5 @@
-using System;
-using System.Collections;
-using System.Reflection.Metadata;
-
 namespace FlashcardApp.Api
 {
-    public class ResetCode
-    {
-        private static Hashtable _codes = new Hashtable();
-        private static Random random = new Random();
-        public static string RandomString(int length)
-        {
-            const string chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
-            return new string(Enumerable.Repeat(chars, length)
-                .Select(s => s[random.Next(s.Length)]).ToArray());
-        }
-        public string GenerateCode (string email)
-        {
-            string code = RandomString(8);
-            try
-            {
-                _codes.Add(email, code);
-            }
-            catch
-            {
-                _codes.Remove(email);
-                _codes.Add(email, code);
-            }
-            Countdown(email);
-            return code;
-        }
-        public bool ValidateResetCode (string email, string code)
-        {
-            if (email == null || code == null)
-            {
-                return false;
-            }
-            
-            if (code == _codes[email].ToString())
-            {
-                return true;
-            }
-            return false;
-        }
-        private void Countdown(string email)
-        {
-            Task.Delay(new TimeSpan(0, 15, 0)).ContinueWith(t => {
-                try
-                {
-                    _codes.Remove(email);
-                }
-                catch
-                {
-
-                }
-            });
-        }
-    }
     public class Program
     {
